gw: add default connection tracking rules even when NodePort is disabled
One can selectively enable NodePort on nodes in K8s cluster. This is
so that OVN control plane can scale well in large cluster. For example:
If you have 1000 node cluster, then you may not want to enable Load
Balancers on all of the 1000 nodes.

We only enable NodePort on few 10s of nodes that form the control plane
services nodes. The rest of the 990 nodes we don’t want it to be a
forwarding or routing entity for these services.

Irrespective of whether NodePort is enabled or not, we still need the
default OpenFlow rules on the host OVS bridge to connection track the
traffic comming in from the OVN logical topology (North-South).

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>