Add/Modify utility methods to cni
`cni.getPodAnnotations` is moved to the new file
and modified to accept a condition function
which will be invoked with the pod's annotation
waiting on a defined condition.

We define two condition methods:

1. ovnReady : check that pod-network annotation is set
   by ovnkube master
2. smartNICReady : check the former, in addition check
   smartnic.connection-status annotation and
   ensure readiness.

Ddd an additional utility method: `PodAnnotation2PodInfo`
This method creates PodInterfaceInfo from Pod annotations.

This method will be later used by both cni and node packages.

While This could have been achieved without creating a new file,
This opens the opportunity to move additional "util-like" functionality
to the new file.

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>