CNI support Smart-NIC
This commit builds on top of the last two commits.
using the new utility methods to implement cmdAdd
and cmdCheck flow for smart-nic.

cmdAdd:
The new flow sets up `smartnic.connection-details`
pod annotation with the needed information to perform
the network plumbing on the Smart-NIC.

ovnkube node running on Smart-NIC then updates
smartnic.connection-status annotation which we wait
on in cmdAdd call.

cmdCheck:
The flow remains similar to what is done today,
with the addition of checking the Pod's
`smartnic.connection-status.status` field

Signed-off-by: Adrian Chiris <adrianc@nvidia.com>
Co-authored-by: Moshe Levi <moshele@nvidia.com>