compare only addresses and ports in endpoint updates callback handler
the current code does a recursive comparison of endpoint.Subset struct
between the old and new endpoints. an endpint address has following
fields

  - ip: 10.0.0.91
    nodeName: ngn-master-02
    targetRef:
      kind: Pod
      name: ovnkube-master-68d9dcb458-59mdg
      namespace: ovn-kubernetes
      resourceVersion: "201250854"
      uid: 251bc3ac-edfc-4944-92a4-dcee469dcb66

it captures the Pod that has the endpoint address. now if that pod is
constantly getting updated, then its resource version changes. say it
is getting updated for other reasons other than pod or the port it
exposes, then the endpoint updates will continue to be sent.

instead we just need to check if the addresses and ports have changed
between the old and new endpoint object

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>