NewAddressSet: return nil in case of error
newOvnAddressSets returns a concrete type, that gets assigned to an
interface. This makes the interface not being nil (even if the backing
value is).  In particular, having
nsInfo.addressSet, err = oc.createNamespaceAddrSetAllPods(ns)
called twice would result in having the interface not being nil but the
value behind it being nil and it will make sequent use of
nsInfo.addressSet refer to a nil value.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>