fix SIGSEGV issue in namespace code
this is an another instance of golang interface holding nil is not
always nil since it holds type too. a golang interface has two values
<type, value>. an interface is nil if both type and value is nil. so,
you cannot directly set the value to an interface and later check if
value was nil by comparing the interface to nil. this is because if the
value is `nil`, then the interface will still hold the type of the
value being set.

with the current code we were hitting SIGSEGV while acquiring as' lock

--------8<-----------8<-----
addPodtoNamespace() {
    <snip>
	if nsInfo.addressSet == nil {   <--- addressSet is really nil here
		nsInfo.addressSet, err = oc.createNamespaceAddrSetAllPods(ns)
		if err != nil {
			return fmt.Errorf("unable to add pod to namespace., "+
				"error: %v", ns, err)
		}
	}

	if err := nsInfo.addressSet.AddIPs\
	     (createIPAddressSlice(portInfo.ips)); err != nil {
	}
	<snip>
}
--------8<-----------8<-----

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>