Obtain all endpoints instead of getting only ready endpoints in endpointslices events.
Currently, we are adding ports to firewalld zones
only if the endpoint moves to ready state and remove it if the endpoint
moves to not ready state. In the scenario of pod restarts,
ports gets removed and added through firewalld commands
as there is continous change of endpoint state from Not-ready to ready.
so, we remove this check of getting ready endpoints .

Signed-off-by: Pardhakeswar Pacha <ppacha@nvidia.com>
(cherry picked from commit 0900d2b2be363029827efb43b0f302c365ec8f86)
Signed-off-by: Yun Zhou <yunz@nvidia.com>