quit retrying pod setup if the pod is already gone from informers cache
we have seen instances where the pod delete event gets queued up and the
pod's DeleteFunc() callback is delayed. in the meantime, we are trying
to do pod setup of the failed pod again and again. since the pod is gone
we fail in setting pod annotations in addLogicalPort(). we add the pod
back to the retryLoop and retry again after 1 minute.

we continue this until the pod's DeleteFunc() gets called and it removes
the deleted pod from the retryPod map.