validate the MTU check against the interface that has OVN Encap IP
whether the MTU is big enough to hold the GENEVE packets is being
erroneously checked against the gatewayInterface (aka bridge interface).
however, that might not be the case in some deployments. it could be that
a separate VTEP network could be used to carry GENEVE packets. so, the
right thing to do is to arrive at the interface from the ovn-encap-ip and
then check for that interface's MTU

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>