Don't override the default value of --monitor-all in dist scripts.
This fixes commit c706d0244e. When the option is not supplied for the
scripts, just let it use default from the binary, so that in the future
when we change the default value only one place needs to be update.

Signed-off-by: Han Zhou <hzhou@ovn.org>
(cherry picked from commit 76b947b6e04c66a52a74710b32bd01cedf01d036)