add support for multiple-BF2 adapters on a single K8s node
when a Pod is configured with a ASAP2 VF, we annotate the pod with the
VF details -- (pfId, vfId) -- which is an index into the PF and VF
within that PF. For example (0, 15) is 0th PF and 15th VF off of that
VF.

in the case of multiple BF2 adapters use case, this will not help. so,
the idea here is to also stash PF's MAC as well. so, the annotation
will look like (pfMAC, pfID, vfID). the ovnkube-node running on the
BF2Adapters now have `shouldSmartNICHandleThisPod()` function which
checks the pfMAC is something that the DPU owns. if it doesn't, then it
just returns doing nothing.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>