don't run shared gateway related code on the non-first BF2 adapter
Consider a K8s node with more than one BF2 adapters. For all the Pods
running on that K8s node, the North-South traffic will be through the
first adapter. Furthermore, the NodePort traffic for the service
exposed by the Pods running on that node is thorugh 1st adapter as well.
So, we will not need to run any of the gateway services on the other
BF2 adapters.

The idea here is to ask the deployment team to set this flag
external_ids:ovn-primary-bf2=false on all the non-primary BF2 adapter.
We read this flag on ovnkube start and skip the gateway part.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>